+++ lib/index.ts
@@ -3,63 +3,75 @@
  * @author huangxueliang
  * @author mcler
  */
-const { Compilation } = require('webpack');
-const { ConcatSource, OriginalSource } = require('webpack-sources');
-const { createHash } = require('crypto');
-const globby = require('globby');
-const upath = require('upath');
-const validateOptions = require('schema-utils').validate;
-const HtmlWebpackPlugin = require('html-webpack-plugin');
+import { Compilation } from 'webpack';
+import type { Compiler, Resolver } from 'webpack';
+import { ConcatSource, OriginalSource } from 'webpack-sources';
+import type { Source } from 'webpack-sources';
+import { createHash } from 'crypto';
+import path from 'path';
+import upath from 'upath';
+import { validate as validateOptions } from 'schema-utils';
+import type { Schema } from 'schema-utils/declarations/validate';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
 
-const schema = require('./schema.json');
-const { createFileWithMap } = require('./file');
-const { ensureTrailingSlash } = require('./utils');
+import schema from './schema.json';
+import { createFileWithMap } from './file';
+import { glob } from './glob';
+// import { ensureTrailingSlash } from './utils';
+
+import type {
+    ConcatPluginInputOptions, ConcatPluginOptions,
+    HtmlWebpackPluginAssets,
+    PromiseResult,
+    Sources, WebpackFileTimestamps,
+} from './types';
 
 const PLUGIN_NAME = 'webpackConcatPlugin';
-
-/** @typedef {import("webpack").Compiler} Compiler */
-/** @typedef {import("webpack").Compilation} Compilation */
-/** @typedef {import("webpack").Context} Context */
-/** @typedef {import("webpack").Resolver} Resolver */
-/** @typedef {import("webpack-sources").Source} Source */
-/** @typedef {Source[]} Sources  */
 
 /**
  * @class ConcatPlugin
  */
-class ConcatPlugin {
-    constructor(optionsArg) {
-        const options = {
+export default class ConcatPlugin {
+    private fileHash: string = '';
+
+    private filesToConcatAbsolutePromise?: Promise<string[]>;
+
+    private finalFileName: string = '';
+
+    private getReadFilePromise?: (_: boolean) => Promise<Sources>;
+
+    private needCreateNewFile: boolean = true;
+
+    private prevFileTimestamps: WebpackFileTimestamps = new Map();
+
+    private resolveCache: Record<string, boolean> = {};
+
+    private settings: ConcatPluginOptions;
+
+    private startTime: number;
+
+    constructor(optionsArg: ConcatPluginInputOptions) {
+        const outputPath = typeof optionsArg.outputPath === 'string' ? optionsArg.outputPath : '';
+        const options: ConcatPluginOptions = {
             fileName: '[name].js',
             name: 'result',
             injectType: 'prepend',
-            outputPath: '',
             ...optionsArg,
+            outputPath,
         };
 
         if (!options.filesToConcat || !options.filesToConcat.length) {
             throw new Error(`${PLUGIN_NAME}: option filesToConcat is required and should not be empty`);
         }
 
-        validateOptions(schema, options, PLUGIN_NAME);
-
-        options.outputPath = options.outputPath && ensureTrailingSlash(options.outputPath);
+        validateOptions(schema as Schema, options);
 
         this.settings = options;
 
-        // used to determine if we should emit files during compiler emit event
         this.startTime = Date.now();
-        this.prevTimestamps = {};
-        this.needCreateNewFile = true;
-        this.resolveCache = {};
     }
 
-    /**
-     * @param {string} fileContent
-     * @param {string} filePath
-     * @returns {string}
-     */
-    getFileName(fileContent, filePath = this.settings.fileName) {
+    private getFileName(fileContent: string, filePath = this.settings.fileName): string {
         if (!this.needCreateNewFile) {
             return this.finalFileName;
         }
@@ -75,29 +87,25 @@ class ConcatPlugin {
             }
 
             const regResult = hashRegExp.exec(filePath);
-            const hashLength = regResult[1] ? Number(regResult[1]) : fileHash.length;
+            const hashLength = (regResult?.[1]) ? Number(regResult[1]) : fileHash.length;
 
             filePath = filePath.replace(hashRegExp, fileHash.slice(0, hashLength));
         }
         return filePath.replace(fileRegExp, this.settings.name);
     }
 
-    /**
-     * @param {string} fileContent
-     * @returns {string}
-     */
-    hashFile(fileContent) {
+    private hashFile(fileContent: string): string {
         if (this.fileHash && !this.needCreateNewFile) {
             return this.fileHash;
         }
 
         const { hashFunction = 'md5', hashDigest = 'hex' } = this.settings;
 
-        this.fileHash = createHash(hashFunction).update(fileContent).digest(hashDigest);
+        let fileHash = createHash(hashFunction).update(fileContent).digest(hashDigest);
 
         if (hashDigest === 'base64') {
             // these are not safe url characters.
-            this.fileHash = this.fileHash.replace(/[/+=]/g, (c) => {
+            fileHash = fileHash.replace(/[/+=]/g, (c) => {
                 switch (c) {
                     case '/': return '_';
                     case '+': return '-';
@@ -107,36 +115,21 @@ class ConcatPlugin {
             });
         }
 
-        return this.fileHash;
+        this.fileHash = fileHash;
+
+        return fileHash;
     }
 
-    /**
-     * @param {Context} context
-     * @returns {Promise<string[]>}
-     */
-    getRelativePathAsync(context) {
-        return Promise.all(this.settings.filesToConcat.map((f) => {
-            if (globby.hasMagic(f)) {
-                return globby(f, {
-                    cwd: context,
-                    nodir: true,
-                });
-            }
-            return f;
-        }))
-            .then((raw) => raw.reduce((target, resource) => target.concat(resource), []))
+    private getRelativePathAsync(context: string): Promise<string[]> {
+        return Promise.all(this.settings.filesToConcat.map((f) => glob(f, context)))
+            .then((resources) => resources.reduce((target: string[], resource) => target.concat(...resource), [] as string[]))
             .catch((error) => {
                 console.error(error);
+                return [] as string[];
             });
     }
 
-    /**
-     * @param {Resolver} resolver
-     * @param {string} context
-     * @param {string} relativeFilePath
-     * @returns {Promise<string>}
-     */
-    resolveReadFile(resolver, context, relativeFilePath) {
+    private resolveReadFile(resolver: Resolver, context: string, relativeFilePath: string): Promise<string> {
         return new Promise((resolve, reject) => {
             resolver.resolve(
                 {},
@@ -148,25 +141,17 @@ class ConcatPlugin {
                         if (!this.resolveCache[relativeFilePath]) reject(error);
                     } else {
                         this.resolveCache[relativeFilePath] = true;
-                        resolve(filePath);
+                        resolve(filePath as string);
                     }
                 },
             );
         });
     }
 
-    /**
-     * @param {Compiler} compiler
-     */
-    resolveReadFiles(compiler) {
+    private resolveReadFiles(compiler: Compiler) {
         const self = this;
 
-        /**
-         * @var {ReturnType<typeof createSourcesPromise>}
-         */
-        let readFilePromise;
-
-        const relativePathArrayPromise = this.getRelativePathAsync(compiler.options.context);
+        const relativePathArrayPromise = this.getRelativePathAsync(compiler.options.context!);
 
         /**
          * @var {Promise<Sources>}
@@ -177,33 +162,34 @@ class ConcatPlugin {
                     .then((relativeFilePathArray) => Promise.all(
                         relativeFilePathArray.map((relativeFilePath) => this.resolveReadFile(
                             resolver,
-                            compiler.options.context,
+                            compiler.options.context!,
                             relativeFilePath,
                         )),
                     )).catch((error) => {
                         console.error(error);
+                        return [];
                     }));
             });
         });
 
-        /**
-         * @returns {Promise<Sources>}
-         */
-        const createNewPromise = () => {
+        const createNewPromise = (): Promise<Sources> => {
             self.needCreateNewFile = true;
 
-            return this.filesToConcatAbsolutePromise
+            return this.filesToConcatAbsolutePromise!
                 .then((filePathArray) => Promise.allSettled(
                     filePathArray.map((filePath) => createFileWithMap(compiler, filePath)),
                 ))
-                .then((results) => results.reduce((sources, { value }) => {
-                    if (value) sources.push(value);
+                .then((results) => results.reduce((sources: Sources, result: PromiseResult<Source>) => {
+                    if (result.status === 'fulfilled') sources.push(result.value);
                     return sources;
                 }, []))
                 .catch((error) => {
                     console.error(error);
+                    return [];
                 });
         };
+
+        let readFilePromise: ReturnType<typeof createNewPromise>;
 
         /**
          * @returns {Promise<Sources>}
@@ -216,11 +202,7 @@ class ConcatPlugin {
         };
     }
 
-    /**
-     * @param {Compilation} compilation
-     * @param {Sources} sources
-     */
-    resolveConcatAndUglify(compilation, sources) {
+    private resolveConcatAndUglify(compilation: Compilation, sources: Sources) {
         const concatSource = new ConcatSource();
         sources.forEach((source, idx) => {
             // New line insertion
@@ -229,24 +211,21 @@ class ConcatPlugin {
                 const currentSourceText = source.source().toString();
                 if (prevSourceText.slice(-1) !== '\n'
                     && currentSourceText.slice(0, 1) !== '\n') {
-                    concatSource.add(new OriginalSource('\n'));
+                    concatSource.add(new OriginalSource('\n', 'x'));
                 }
             }
             concatSource.add(source);
         });
         this.finalFileName = this.getFileName(concatSource.source().toString());
         compilation.emitAsset(
-            upath.join(this.settings.outputPath, this.finalFileName),
-            concatSource,
+            path.join(this.settings.outputPath, this.finalFileName),
+            concatSource as any, // Reason for any: own Source typing in webpack
         );
 
         this.needCreateNewFile = false;
     }
 
-    /**
-     * @param {Compiler} compiler
-     */
-    apply(compiler) {
+    apply(compiler: Compiler) {
         // ensure only compile one time per emit
         let compileLoopStarted = false;
 
@@ -254,79 +233,92 @@ class ConcatPlugin {
 
         const self = this;
 
-        const dependenciesChanged = (compilation, filesToConcatAbsolute) => {
-            if (!compilation.fileTimestamps) {
+        const dependenciesChanged = (compilation: Compilation, filesToConcatAbsolute: string[]) => {
+            // Reason for any: non-full Compilation typing
+            const fileTimestamps: WebpackFileTimestamps = (compilation as any).fileTimestamps as WebpackFileTimestamps;
+            if (!fileTimestamps) {
                 return true;
             }
-            const fileTimestampsKeys = Array.from(compilation.fileTimestamps.keys());
+            const fileTimestampsKeys = Array.from(fileTimestamps.keys());
             if (!fileTimestampsKeys.length) {
                 return true;
             }
             const changedFiles = fileTimestampsKeys.filter(
                 (file) => {
-                    const start = this.prevTimestamps.get(file) || this.startTime;
-                    const end = compilation.fileTimestamps.get(file) || Infinity;
+                    const start = this.prevFileTimestamps.get(file) || this.startTime;
+                    const end = fileTimestamps.get(file) || Infinity;
                     return start < end;
                 },
             );
 
-            this.prevTimestamps = compilation.fileTimestamps;
+            this.prevFileTimestamps = fileTimestamps;
 
             return changedFiles.some((file) => filesToConcatAbsolute.includes(file));
         };
 
-        /**
-         * @param {Compilation} compilation
-         * @param {Function} callback
-         * @returns {void}
-         */
-        const processCompiling = (compilation, callback) => {
-            self.filesToConcatAbsolutePromise.then((filesToConcatAbsolute) => {
+        const processCompiling = (compilation: Compilation, callback: Function) => {
+            self.filesToConcatAbsolutePromise!.then((filesToConcatAbsolute) => {
                 filesToConcatAbsolute.forEach((file) => {
-                    compilation.fileDependencies.add(upath.relative(compiler.options.context, file));
+                    compilation.fileDependencies.add(upath.relative(compiler.options.context!, file));
                 });
                 if (!dependenciesChanged(compilation, filesToConcatAbsolute)) {
                     return callback();
                 }
-                return self.getReadFilePromise(true).then((files) => {
+                return self.getReadFilePromise!(true).then((files) => {
                     self.resolveConcatAndUglify(compilation, files);
 
                     callback();
                 });
             }).catch((error) => {
                 console.error(error);
+                // callback();
             });
         };
 
         compiler.hooks.compilation.tap(PLUGIN_NAME, (compilation) => {
-            let assetPath;
+            let assetPath: string;
             let hookBeforeAssetTagGeneration = HtmlWebpackPlugin.getHooks(compilation).beforeAssetTagGeneration;
 
             hookBeforeAssetTagGeneration && hookBeforeAssetTagGeneration.tapAsync(PLUGIN_NAME, (htmlPluginData, callback) => {
                 const getAssetPath = () => {
                     if (typeof self.settings.publicPath === 'undefined') {
                         if (typeof htmlPluginData.assets.publicPath === 'undefined') {
-                            return upath.relative(upath.dirname(htmlPluginData.outputName), `${self.settings.outputPath}${self.finalFileName}`);
+                            return path.relative(
+                                path.dirname(htmlPluginData.outputName),
+                                path.join(self.settings.outputPath, self.finalFileName),
+                            );
                         }
-                        return `${ensureTrailingSlash(htmlPluginData.assets.publicPath)}${self.settings.outputPath}${self.finalFileName}`;
+                        return path.join(
+                            htmlPluginData.assets.publicPath,
+                            self.settings.outputPath,
+                            self.finalFileName,
+                        );
                     }
-                    if (self.settings.publicPath === false) {
-                        return upath.relative(upath.dirname(htmlPluginData.outputName), `${self.settings.outputPath}${self.finalFileName}`);
+                    if (!self.settings.publicPath) {
+                        return path.relative(
+                            path.dirname(htmlPluginData.outputName),
+                            path.join(self.settings.outputPath, self.finalFileName),
+                        );
                     }
-                    return `${ensureTrailingSlash(self.settings.publicPath)}${self.settings.outputPath}${self.finalFileName}`;
+                    return path.join(
+                        self.settings.publicPath,
+                        self.settings.outputPath,
+                        self.finalFileName,
+                    );
                 };
 
                 const injectToHtml = () => {
-                    htmlPluginData.assets.webpackConcat = htmlPluginData.assets.webpackConcat || {};
+                    const htmlWebpackPluginAssets: HtmlWebpackPluginAssets = htmlPluginData.assets as HtmlWebpackPluginAssets;
+                    if (!htmlWebpackPluginAssets.webpackConcat) htmlWebpackPluginAssets.webpackConcat = {};
 
                     assetPath = getAssetPath();
 
-                    htmlPluginData.assets.webpackConcat[self.settings.name] = assetPath;
+                    htmlWebpackPluginAssets.webpackConcat[self.settings.name] = assetPath;
 
                     if (self.settings.injectType === 'prepend') {
-                        htmlPluginData.assets.js.unshift(assetPath);
+                        htmlWebpackPluginAssets.js.unshift(assetPath);
                     } else if (self.settings.injectType === 'append') {
-                        htmlPluginData.assets.js.push(assetPath);
+                        htmlWebpackPluginAssets.js.push(assetPath);
                     }
                 };
 
@@ -357,11 +349,11 @@ class ConcatPlugin {
             });
         });
 
-        compiler.hooks.thisCompilation.tap(PLUGIN_NAME, (compilation) => {
+        compiler.hooks.thisCompilation.tap(PLUGIN_NAME, (compilation: Compilation) => {
             compilation.hooks.processAssets.tapAsync({
                 name: PLUGIN_NAME,
                 stage: Compilation.PROCESS_ASSETS_STAGE_ADDITIONAL,
-            }, (_, callback) => {
+            }, (_, callback: Function) => {
                 if (!compileLoopStarted) {
                     compileLoopStarted = true;
                     processCompiling(compilation, callback);
@@ -375,5 +367,3 @@ class ConcatPlugin {
         });
     }
 }
-
-module.exports = ConcatPlugin;